Rename default ProgressBar to TQDMProgressBar
