+++ pytorch_lightning/callbacks/progress/__init__.py
@@ -19,5 +19,6 @@ Use or override one of the progress bar callbacks.
 
 """
 from pytorch_lightning.callbacks.progress.base import ProgressBarBase  # noqa: F401
-from pytorch_lightning.callbacks.progress.progress import ProgressBar, tqdm  # noqa: F401
+from pytorch_lightning.callbacks.progress.progress import ProgressBar  # noqa: F401
 from pytorch_lightning.callbacks.progress.rich_progress import RichProgressBar  # noqa: F401
+from pytorch_lightning.callbacks.progress.tqdm_progress import tqdm, TQDMProgressBar  # noqa: F401
